feat : enter 키 이벤트. 클릭 시 바로 검색
